+++ src-gui/src/renderer/components/pages/help/DonateInfoBox.tsx
@@ -13,22 +13,8 @@ export default function DonateInfoBox() {
       icon={<MoneroIcon />}
       additionalContent={
         <Typography variant="subtitle2">
-          <p>
-            As part of the Monero Community Crowdfunding System (CCS), we
-            received funding for 6 months of full-time development by generous
-            donors from the Monero community (
-            <Link
-              href="https://ccs.getmonero.org/proposals/mature-atomic-swaps-ecosystem.html"
-              target="_blank"
-            >
-              link
-            </Link>
-            ).
-          </p>
-          <p>
-            If you want to support our effort event further, you can do so at
-            this address.
-          </p>
+          If you want to support our effort event further, you can do so at this
+          address.
         </Typography>
       }
     />