+++ src-gui/src/renderer/components/pages/swap/swap/init/InitPage.tsx
@@ -1,125 +1,85 @@
-import { Box, Paper, Tab, Tabs, Typography } from "@mui/material";
-import PlayArrowIcon from "@mui/icons-material/PlayArrow";
-import { useState } from "react";
-import BitcoinAddressTextField from "renderer/components/inputs/BitcoinAddressTextField";
-import MoneroAddressTextField from "renderer/components/inputs/MoneroAddressTextField";
-import PromiseInvokeButton from "renderer/components/PromiseInvokeButton";
+import { Box, Button, Skeleton } from "@mui/material";
 import { buyXmr } from "renderer/rpc";
-import { useSettings } from "store/hooks";
+import { useSnackbar } from "notistack";
+import SearchIcon from "@mui/icons-material/Search";
 
 export default function InitPage() {
-  const [redeemAddress, setRedeemAddress] = useState("");
-  const [refundAddress, setRefundAddress] = useState("");
-  const [useExternalRefundAddress, setUseExternalRefundAddress] =
-    useState(false);
-
-  // We force this to true for now because the internal wallet is not really accessible from the GUI yet
-  const [useExternalRedeemAddress, setUseExternalRedeemAddress] =
-    useState(true);
-
-  const [redeemAddressValid, setRedeemAddressValid] = useState(false);
-  const [refundAddressValid, setRefundAddressValid] = useState(false);
-
-  const donationRatio = useSettings((s) => s.donateToDevelopment);
+  const { enqueueSnackbar } = useSnackbar();
 
   async function init() {
-    await buyXmr(
-      useExternalRefundAddress ? refundAddress : null,
-      useExternalRedeemAddress ? redeemAddress : null,
-      donationRatio,
-    );
+    try {
+      await buyXmr();
+    } catch (error) {
+      enqueueSnackbar(error as string, {
+        variant: "error",
+      });
+    }
   }
 
   return (
-    <>
+    <Box
+      sx={{
+        display: "flex",
+        flexDirection: "column",
+        gap: 2,
+        flex: 1,
+        position: "relative",
+        cursor: "pointer",
+      }}
+      onClick={init}
+    >
+      <Skeleton variant="rounded" sx={{ flex: 1, minHeight: "20vh" }} />
       <Box
         sx={{
+          flex: 1,
           display: "flex",
           flexDirection: "column",
-          gap: 1.5,
+          gap: 2,
+          borderRadius: 1,
         }}
       >
-        <Paper variant="outlined" style={{}}>
-          <Tabs
-            value={useExternalRedeemAddress ? 1 : 0}
-            indicatorColor="primary"
-            variant="fullWidth"
-            onChange={(_, newValue) =>
-              setUseExternalRedeemAddress(newValue === 1)
-            }
+        {[...Array(4)].map((_, index) => (
+          <Box
+            sx={{ display: "flex", alignItems: "center", gap: 2 }}
+            key={index}
           >
-            <Tab label="Redeem to internal Monero wallet" value={0} />
-            <Tab label="Redeem to external Monero address" value={1} />
-          </Tabs>
-          <Box style={{ padding: "16px" }}>
-            {useExternalRedeemAddress ? (
-              <MoneroAddressTextField
-                label="External Monero redeem address"
-                address={redeemAddress}
-                onAddressChange={setRedeemAddress}
-                onAddressValidityChange={setRedeemAddressValid}
-                helperText="The monero will be sent to this address if the swap is successful."
-                fullWidth
-              />
-            ) : (
-              <Typography variant="caption">
-                The Monero will be sent to the internal Monero wallet of the
-                GUI. You can then withdraw them from there or use them for
-                another swap directly.
-              </Typography>
-            )}
+            <Skeleton variant="circular" width={40} height={40} />
+            <Skeleton
+              variant="rounded"
+              animation="wave"
+              height={"7vh"}
+              sx={{ flex: 1 }}
+            />
           </Box>
-        </Paper>
+        ))}
+      </Box>
 
-        <Paper variant="outlined" style={{}}>
-          <Tabs
-            value={useExternalRefundAddress ? 1 : 0}
-            indicatorColor="primary"
-            variant="fullWidth"
-            onChange={(_, newValue) =>
-              setUseExternalRefundAddress(newValue === 1)
-            }
-          >
-            <Tab label="Refund to internal Bitcoin wallet" value={0} />
-            <Tab label="Refund to external Bitcoin address" value={1} />
-          </Tabs>
-          <Box style={{ padding: "16px" }}>
-            {useExternalRefundAddress ? (
-              <BitcoinAddressTextField
-                label="External Bitcoin refund address"
-                address={refundAddress}
-                onAddressChange={setRefundAddress}
-                onAddressValidityChange={setRefundAddressValid}
-                helperText="In case something goes wrong, the Bitcoin will be refunded to this address."
-                fullWidth
-              />
-            ) : (
-              <Typography variant="caption">
-                In case something goes wrong, the Bitcoin will be refunded to
-                the internal Bitcoin wallet of the GUI. You can then withdraw
-                them from there or use them for another swap directly.
-              </Typography>
-            )}
-          </Box>
-        </Paper>
-      </Box>
-      <Box style={{ display: "flex", justifyContent: "center" }}>
-        <PromiseInvokeButton
-          disabled={
-            (!refundAddressValid && useExternalRefundAddress) ||
-            (!redeemAddressValid && useExternalRedeemAddress)
-          }
+      <Box
+        sx={{
+          position: "absolute",
+          top: 0,
+          left: 0,
+          right: 0,
+          bottom: 0,
+          display: "flex",
+          alignItems: "flex-start",
+          justifyContent: "center",
+          paddingTop: "15%",
+          margin: "-1rem",
+          borderRadius: "1rem",
+          pointerEvents: "none",
+          backdropFilter: "blur(1px)",
+        }}
+      >
+        <Button
           variant="contained"
-          color="primary"
           size="large"
-          sx={{ marginTop: 1 }}
-          endIcon={<PlayArrowIcon />}
-          onInvoke={init}
-          displayErrorSnackbar
+          sx={{ pointerEvents: "auto" }}
+          endIcon={<SearchIcon />}
         >
-          Continue
-        </PromiseInvokeButton>
+          Click to view offers
+        </Button>
       </Box>
-    </>
+    </Box>
   );
 }