+++ swap/src/cli/api/tauri_bindings.rs
@@ -873,7 +873,6 @@ pub struct TauriSwapProgressEventWrapper {
 #[typeshare]
 #[serde(tag = "type", content = "content")]
 pub enum TauriSwapProgressEvent {
-    RequestingQuote,
     Resuming,
     ReceivedQuote(BidQuote),
     WaitingForBtcDeposit {