+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GUI + SWAP + ASB: Upgrade arti (tor library) to 1.5.0. This might improve connectivity reliability.
 - ASB: Fix an issue where we would not wait between re-dials of rendezvous nodes.
 - GUI: Faster startup time by allowing parts of the application be used while other components are still initializing.
+- GUI: We now default to redeeming swaps into the internal Monero wallet, and sending Bitcoin refund into the internal Bitcoin wallet. If you want to change this behaviour go to Settings -> "Redeem Policy" and "Refund Policy".
 
 ## [3.0.4] - 2025-09-26